Better organized API now
