+++ src/main/java/com/alexander/syndicatefighter/Skills/FixTheBooksSkill.java
@@ -1,4 +1,6 @@
-package com.alexander.syndicatefighter;
+package com.alexander.syndicatefighter.Skills;
+
+import com.alexander.syndicatefighter.Worker.Worker;
 
 /**
  * Created by alexandernohe on 11/14/15.
