+++ src/main/java/com/alexander/syndicatefighter/Skills/LitigateSkill.java
@@ -1,6 +1,6 @@
-package com.alexander.syndicatefighter;
+package com.alexander.syndicatefighter.Skills;
 
-import com.alexander.syndicatefighter.Skills;
+import com.alexander.syndicatefighter.Worker.Worker;
 
 /**
  * Created by alexandernohe on 11/10/15.
