+++ src/main/java/com/alexander/syndicatefighter/Skills/Skills.java
@@ -1,4 +1,6 @@
-package com.alexander.syndicatefighter;
+package com.alexander.syndicatefighter.Skills;
+
+import com.alexander.syndicatefighter.Worker.Worker;
 
 /**
  * Created by alex7370 on 11/9/2015.
@@ -19,7 +21,7 @@ abstract public class Skills {
         this.costNRG = CostPerAttach;
 
         // When writing subclasses, use the following format of constructor
-        // class SubClass extends com.alexander.syndicatefighter.Skills
+        // class SubClass extends com.alexander.syndicatefighter.Skills.Skills
         // {
         //    public SubClass(String Name, int MaxNRG, int Cooldown, int CostPerAttach)
         //     {   super(String Name, int MaxNRG, int Cooldown, int CostPerAttach);   }
