+++ src/main/java/com/alexander/syndicatefighter/Worker/TeamType.java
@@ -1,0 +1,15 @@
+package com.alexander.syndicatefighter.Worker;
+
+/**
+ * Created by alex7370 on 11/25/2015.
+ */
+public enum TeamType {
+    DEV,
+    HR,
+    MANAGEMENT,
+    FINANCIAL,
+    DATAENTRY,
+    ANALYST,
+    IT,
+    LEGAL
+}
