+++ src/test/java/com/alexander/syndicatefighter/WorkerTest.java
@@ -1,8 +1,9 @@
 package com.alexander.syndicatefighter;
 
-import com.alexander.syndicatefighter.LitigateSkill;
-import com.alexander.syndicatefighter.Skills;
-import com.alexander.syndicatefighter.Worker;
+import com.alexander.syndicatefighter.Skills.Skills;
+import com.alexander.syndicatefighter.Skills.LitigateSkill;
+import com.alexander.syndicatefighter.Worker.Worker;
+import com.alexander.syndicatefighter.Worker.TeamType;
 
 import static org.junit.Assert.*;
 
@@ -45,7 +46,7 @@ public class WorkerTest {
 
     @org.junit.Test
     public void testGetType() throws Exception {
-        assertEquals(Worker.TeamType.DEV, worker.getType());
+        assertEquals(TeamType.DEV, worker.getType());
     }
 
     @org.junit.Test
@@ -67,7 +68,7 @@ public class WorkerTest {
     public void testConstructors() throws Exception {
         Worker worker1 = new Worker("Kwasi");
         assertEquals("Kwasi", worker1.getName());
-        assertEquals(Worker.TeamType.DEV,worker1.getType());
+        assertEquals(TeamType.DEV,worker1.getType());
         assertEquals(1, worker1.getLevel());
         assertEquals(25, worker1.getCurrentXP());
         assertEquals(25, worker1.getMaxHP());
@@ -77,7 +78,7 @@ public class WorkerTest {
         assertEquals("Kwasi", worker2.getName());
         assertEquals(1098903, worker2.getCurrentXP()); //Checked that this should give us level 100 but enforced a level cap.
         assertEquals(99, worker2.getLevel());
-        assertEquals(Worker.TeamType.DEV,worker2.getType());
+        assertEquals(TeamType.DEV,worker2.getType());
         assertEquals(9923, worker2.getMaxHP());  //Checked this manually
         assertEquals(9923, worker2.getCurrentHP()); //Checked this manually
     }
